output: only use the last connected output
Multiple outputs has never really fit with Cage's window layouting model
and had several limitations. It also is difficult to unify with layer
shell surfaces, which expect to be run on specific outputs.

Until we find a way to improve the interaction between layer shell and
multiple outputs, using only the last connected output is a nice
middle-ground.